+++ gala/potential/potential/builtin/exp_fields.cc
@@ -175,8 +175,14 @@ double exp_value(double t, double *pars, double *q, int n_dim, void* state) {
   gala_exp::State *exp_state = static_cast<gala_exp::State *>(state);
 
   if (!exp_state->is_static) {
-    // TODO: how expensive is this, actually?
-    exp_state->basis->set_coefs(gala_exp::interpolator(t, exp_state->coefs));
+    try {
+      // TODO: how expensive is this, actually?
+      exp_state->basis->set_coefs(gala_exp::interpolator(t, exp_state->coefs));
+    } catch (const std::exception &e) {
+      // TODO: propagate this exception to Python
+      std::cerr << "[GALA EXP] Error setting coefficients: " << e.what() << std::endl;
+      return std::numeric_limits<double>::quiet_NaN();
+    }
   }
 
   // Get the field quantities
@@ -191,7 +197,14 @@ void exp_gradient(double t, double *pars, double *q, int n_dim, double *grad, vo
   gala_exp::State *exp_state = static_cast<gala_exp::State *>(state);
 
   if (!exp_state->is_static) {
-    exp_state->basis->set_coefs(gala_exp::interpolator(t, exp_state->coefs));
+    try {
+      exp_state->basis->set_coefs(gala_exp::interpolator(t, exp_state->coefs));
+    } catch (const std::exception &e) {
+      // TODO
+      std::cerr << "[GALA EXP] Error setting coefficients: " << e.what() << std::endl;
+      grad[0] = grad[1] = grad[2] = std::numeric_limits<double>::quiet_NaN();
+      return;
+    }
   }
 
   // TODO: ask Martin/Mike for a way to compute only the force/acceleration - we're wasting
@@ -207,7 +220,13 @@ double exp_density(double t, double *pars, double *q, int n_dim, void* state) {
   gala_exp::State *exp_state = static_cast<gala_exp::State *>(state);
 
   if (!exp_state->is_static) {
-    exp_state->basis->set_coefs(gala_exp::interpolator(t, exp_state->coefs));
+    try {
+      exp_state->basis->set_coefs(gala_exp::interpolator(t, exp_state->coefs));
+    } catch (const std::exception &e) {
+      // TODO
+      std::cerr << "[GALA EXP] Error setting coefficients: " << e.what() << std::endl;
+      return std::numeric_limits<double>::quiet_NaN();
+    }
   }
 
   // TODO: ask Martin/Mike for a way to compute only the density - we're wasting